renaming marketing section
