+++ dataland-frontend/src/components/pages/WelcomeDataland.vue
@@ -7,7 +7,7 @@
 
 <script>
 import LandingLogin from "@/components/resources/landing/LandingLogin.vue"
-import SampleSection from "@/components/resources/landing/Marketing";
+import SampleSection from "@/components/resources/landing/MarketingSection";
 import MarketingSection from "@/components/resources/landing/SampleSection";
 import LandingFooter from "@/components/resources/landing/LandingFooter";
 export default {
