+++ dataland-frontend/src/components/resources/landing/MarketingSection.vue
